Simplify index.html and style.css classes
Replace class of each specific section name to 'container' and removed
unneccesary classes from both index.html and style.css. Rename certain
classes to be more "reusable" as a class.

Before each starting div was named specific to the section. This caused
the code to be hard to read and caused complications when it came to
editing classes.

This change should make the website more uniform and increase work flow